GitHubProjects_RRR_V5/road-freight-risk-ai-v3-main/backend/requirements.txt
# ✅ Core API
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
starlette==0.36.3
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
PyJWT==2.8.0

# ✅ Email & Security
python-multipart==0.0.6
aiofiles==23.2.1

# ✅ ML/NLP
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
textblob==0.18.0.post0
nltk==3.8.1

# ✅ HTTP Requests
requests==2.31.0
httpx==0.25.2

# ✅ Map packages - Compatible with folium 0.14.0
folium==0.14.0
streamlit-folium==0.16.0
branca==0.6.0  # Required for folium 0.14.0 color mapping

# ✅ Voice Processing
SpeechRecognition==3.10.0
pydub==0.25.1
langdetect==1.0.9

# ✅ File Processing
Pillow==10.0.1
python-magic==0.4.27

# ✅ Database (optional - for production)
# sqlalchemy==2.0.23
# alembic==1.12.1

# ✅ Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# ✅ Compatibility notes:
# - folium 0.14.0 is stable and widely compatible
# - streamlit-folium 0.16.0 works well with folium 0.14.0
# - branca 0.6.0 is required for folium 0.14.0 color mapping
